+++ src/university/UI/StudentManagementScreen.java
@@ -2,12 +2,10 @@ package university.UI;
 
 import java.awt.*;
 import javax.swing.*;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JButton;
-import javax.swing.SwingConstants;
 
 import university.UI.ProfileScreen;
+import university.Account;
+import university.Module;
 import university.ScreenManager;
 import university.Student;
 import university.TableModel;
@@ -70,14 +68,34 @@ class StudentManagementScreen extends JPanel implements ActionListener {
                 int index = Integer.parseInt((String) studentTable.getValueAt(studentTable.getSelectedRow(), 0));
                 String username = (String) studentTable.getValueAt(studentTable.getSelectedRow(), 3);
                 Student stuToDelete = new Student(index);
-                try {
-                    stuToDelete.deleteStudent();
-                } catch (SQLException e1) {
-                    e1.printStackTrace();
+                
+                
+                
+                JLabel label_login = new JLabel("If you wish to continue please insert your account details.");
+                JLabel labem_email = new JLabel("Email:");
+                JTextField email = new JTextField();
+
+                JLabel label_password = new JLabel("Password:");
+                JPasswordField password = new JPasswordField();
+
+                Object[] array = {label_login, labem_email, email, label_password, password };
+
+                int res = JOptionPane.showConfirmDialog(null, array, "Login", JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+                if (res == JOptionPane.OK_OPTION) {
+	                try {
+	                	if(Account.delVerification(email.getText(), password.getText(), 3)) {
+	                		stuToDelete.deleteStudent();
+	                		
+	                	}else
+	                		JOptionPane.showMessageDialog(null, "Please insert the correct account details");
+                    
+	                } catch (SQLException e1) {
+	                    e1.printStackTrace();
+	                }
+	                this.studentManagement.setVisible(false);
+	                this.profileScreen.draw();
+	                JOptionPane.showMessageDialog(null, "Successfully deleted student: " + username);
                 }
-                this.studentManagement.setVisible(false);
-                this.profileScreen.draw();
-                JOptionPane.showMessageDialog(null, "Successfully deleted student: " + username);
             } else {
                 JOptionPane.showMessageDialog(null, "Please select a Student to delete");
             }
